Update trabalho av2(2).py

+++ trabalho av2(2).py
@@ -1,5 +1,5 @@
 valor_produto = float(input('Valor do produto: '))
-taxa_imposto = int(input('Quanto de imposto: '))
+taxa_imposto = float(input('Quanto de imposto: '))
 por_taxa = taxa_imposto / 100
 porcentagem_imposto = valor_produto * por_taxa
 print(f'Valor do produto antes: {valor_produto}')
